+++ quilt/src/main/java/de/maxhenkel/voicechat/QuiltVoicechatClientMod.java
@@ -1,5 +1,6 @@
 package de.maxhenkel.voicechat;
 
+import com.sun.jna.Platform;
 import de.maxhenkel.configbuilder.ConfigBuilder;
 import de.maxhenkel.voicechat.config.QuiltClientConfig;
 import de.maxhenkel.voicechat.integration.ClothConfig;
@@ -14,5 +15,9 @@ public class QuiltVoicechatClientMod extends VoicechatClient implements ClientMo
         CLIENT_CONFIG = ConfigBuilder.build(Minecraft.getInstance().gameDirectory.toPath().resolve("config").resolve(Voicechat.MODID).resolve("voicechat-client.properties"), true, QuiltClientConfig::new);
         initializeClient();
         ClothConfig.init();
+
+        if (Platform.isMac() && !CLIENT_CONFIG.javaMicrophoneImplementation.get()) {
+            CLIENT_CONFIG.javaMicrophoneImplementation.set(true).save();
+        }
     }
 }